risk_explainer/requirements.txt
pandas>=1.3.0
numpy>=1.21.0
xgboost>=1.6.0
shap>=0.41.0
azure-identity>=1.12.0
tabulate>=0.8.9
jupyter>=1.0.0
scikit-learn>=1.0.0
jupyterlab
matplotlib
python-dotenv
openai>=1.24.1